Ensure sharedTreeAttributes can be imported without pulling in lots of other stuff
