+++ examples/utils/bundle-size-tests/src/sharedTreeAttributes.ts
@@ -1,0 +1,11 @@
+/*!
+ * Copyright (c) Microsoft Corporation and contributors. All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+// eslint-disable-next-line import/no-internal-modules
+import { SharedTreeAttributes } from "@fluidframework/tree/legacy";
+
+export function apisToBundle() {
+	return SharedTreeAttributes;
+}
